refactor: move `<Empty>` to `components/empty`
I will be removing all of the other atoms and this is the only one that
is currently being used anywhere.
