+++ app/routes/shows._index.tsx
@@ -1,7 +1,7 @@
 import { Link, useLoaderData } from '@remix-run/react'
 import { useMemo } from 'react'
 
-import { Empty } from 'atoms/Empty'
+import { Empty } from 'components/empty'
 
 import { prisma } from 'db.server'
 import { log } from 'log.server'
